Moved code samples from bin to lib and test
- Moved code samples to the test module
- Moved code that is part of exercises to the lib folder (to be used in
the future by the examples/exercises on dartpads)
