--- examples/iterables/bin/firstwhere.dart
@@ -1,6 +1,0 @@
-String firstWhere(Iterable<String> iterable) {
-  // #docregion firstwhere
-  String element = iterable.firstWhere((element) => element.length > 5);
-  // #enddocregion firstwhere
-  return element;
-}
