--- examples/iterables/bin/iterable_1.dart
@@ -1,1 +1,0 @@
-Iterable<int> iterable = [1, 2, 3];
