--- examples/iterables/bin/iterable_2.dart
@@ -1,2 +1,0 @@
-Iterable<int> iterable = [1, 2, 3];
-int value = iterable.elementAt(1);
