+++ examples/iterables/lib/email.dart
